requirements.txt
openai>=0.27.0
easyocr>=1.6.2
pytesseract>=0.3.10
pillow>=9.0.0
pyyaml>=6.0
mss>=6.1.0
opencv-python>=4.5.5.64
numpy>=1.19.0
